+++ client/src/components/Common/PublishedItem.vue
