+++ client/src/components/Visualizations/VisualizationPublished.vue
@@ -1,20 +1,21 @@
 <template>
-    <Published :item="visualization">
+    <PublishedItem :item="visualization">
         <template v-slot>
             <CenterFrame :src="getUrl" />
         </template>
-    </Published>
+    </PublishedItem>
 </template>
 
 <script>
-import Published from "components/Common/Published";
-import CenterFrame from "entry/analysis/modules/CenterFrame";
-import { urlData } from "utils/url";
+import { urlData } from "@/utils/url";
+
+import PublishedItem from "@/components/Common/PublishedItem.vue";
+import CenterFrame from "@/entry/analysis/modules/CenterFrame.vue";
 
 export default {
     components: {
         CenterFrame,
-        Published,
+        PublishedItem,
     },
     props: {
         id: {
